Don't delete docset if we're running locally

+++ app.rb
@@ -371,7 +371,7 @@ class CocoaDocs < Object
       if $delete_source_after_docset_creation
         vputs "Deleting source files"
         command "rm -rf #{download_location}"
-        command "rm -rf #{docset_location}"
+        command "rm -rf #{docset_location}" if $upload_site_to_s3
       end
     end
